Imported the fixtures to Media1155.test.ts
